+++ compositions/README.md
@@ -1,0 +1,1 @@
+Place Reason files and composition definition JSON files in this directory
