+++ packages/backend/.env.example
@@ -18,7 +18,8 @@ ZAPPERMENT_BACKEND_SCENE_QUALITY=0.7
 # number of random scenes Zapperment attempts to create to match the required scene quality; integer
 ZAPPERMENT_BACKEND_MAX_ATTEMPTS=100
 
-# the track that is playing, name corresponds to that of JSON and Reason files stored in the
-# tracks directory – demo tracks are available for download here:
-# https://github.com/zapperment/zapperment/wiki/Demo-Tracks
-ZAPPERMENT_BACKEND_TRACK=fuzzy-tendrils
+# the composition that is playing, name corresponds to that of JSON and Reason files stored in the
+# compositions directory – demo compositions are available for download here:
+# https://github.com/zapperment/zapperment/wiki/Demo-Compositions
+ZAPPERMENT_BACKEND_COMPOSITION=fuzzy-tendrils
+
