+++ packages/backend/src/composition/buildRandomScene.js
