+++ packages/backend/src/composition/convertValuesToMidiNumbers.js
