+++ packages/backend/src/composition/utils/getCompositionPath.js
@@ -1,0 +1,4 @@
+const { resolve } = require("path");
+
+module.exports = composition =>
+  `${resolve(__dirname, "../../../../../compositions", composition)}.json`;
