+++ packages/backend/src/composition/utils/getMidiNote.js
