+++ packages/backend/src/composition/utils/initErrorInfo.js
