+++ packages/backend/src/composition/utils/interpolateMidiValue.js
