+++ packages/backend/src/composition/utils/isControlledByRotary.js
