+++ packages/backend/src/composition/utils/isNote.js
