+++ packages/backend/src/composition/utils/printErrorInfo.js
