+++ packages/backend/src/composition/utils/processChannel.js
