+++ packages/backend/src/composition/utils/string.js
