+++ packages/backend/src/composition/utils/walk.js
