+++ packages/backend/src/composition/validateAndSetDefaults.js
