--- packages/backend/src/track/utils/getTrackPath.js
@@ -1,4 +1,0 @@
-const { resolve } = require("path");
-
-module.exports = track =>
-  `${resolve(__dirname, "../../../../../tracks", track)}.json`;
