--- tracks/README.md
@@ -1,1 +1,0 @@
-Place Reason files and track definition JSON files in this directory
